backend/requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
requests
pydantic

